+++ packages/sources/vue/src/components/index.ts
@@ -6,6 +6,9 @@ import VtmnBadge from './indicators/VtmnBadge';
 import VtmnLoader from './indicators/VtmnLoader';
 import VtmnPrice from './indicators/VtmnPrice';
 import VtmnTag from './indicators/VtmnTag';
+import VtmnAlert from './overlays/VtmnAlert';
+import VtmnPopover from './overlays/VtmnPopover';
+import VtmnTooltip from './overlays/VtmnTooltip';
 import VtmnCheckbox from './selection-controls/VtmnCheckbox';
 import VtmnChip from './selection-controls/VtmnChip';
 import VtmnRadioButton from './selection-controls/VtmnRadioButton';
@@ -26,6 +29,9 @@ export {
   VtmnChip,
   VtmnRadioButton,
   VtmnTag,
+  VtmnAlert,
+  VtmnPopover,
+  VtmnTooltip,
   VtmnToggle,
   VtmnAccordion,
   VtmnDivider,