+++ packages/sources/vue/src/components/overlays/VtmnPopover/types.ts
@@ -1,0 +1,9 @@
+export type VtmnPopoverPosition =
+  | 'top-left'
+  | 'top'
+  | 'top-right'
+  | 'right'
+  | 'bottom-right'
+  | 'bottom'
+  | 'bottom-left'
+  | 'left';